Removed ComponentParameterBuilder, added support for unmatched and cascading values
