+++ src/bunit.web/TestContext.cs
@@ -33,7 +33,7 @@ namespace Bunit
 		/// <typeparam name="TComponent">Type of the component to render</typeparam>
 		/// <param name="parameterBuilder">The ComponentParameterBuilder action to add type safe parameters to pass to the component when it is rendered</param>
 		/// <returns>The rendered <typeparamref name="TComponent"/></returns>
-		public virtual IRenderedComponent<TComponent> RenderComponent<TComponent>(Action<ComponentParameterBuilder<TComponent>> parameterBuilder) where TComponent : IComponent
+		public virtual IRenderedComponent<TComponent> RenderComponent<TComponent>(Action<ComponentParameterCollectionBuilder<TComponent>> parameterBuilder) where TComponent : IComponent
 			=> TestRendererExtensions.RenderComponent<TComponent>(Renderer, parameterBuilder);
 	}
 }
